requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
clarifai-grpc==7.11.0
emoji==1.7.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
idna==3.3
protobuf==3.20.0
PySocks==1.7.1
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
